requirements.txt
transformers>=4.44.0
accelerate>=0.34.0
torch>=2.4.0
pandas>=2.2.2
pyarrow>=17.0.0
requests>=2.31.0
pyyaml>=6.0.0

